refactor(app): extract HardwareGlue to runtime/hardware_glue.py (#73)
Move the 11-method connector + serial hardware-wiring cluster out of
Runtime into a new HardwareGlue collaborator in runtime/hardware_glue.py,
following the BoatSetup (#72) / DeviceManager (#70) / DepthService (#71)
pattern (self._rt back-reference, imports, construct+delegate).

Methods moved verbatim:
  list_serial_ports, _hw_probe_serial, _build_serial_gps,
  _build_serial_compass, _build_serial_motor, _make_connector_sink,
  connector_status, set_connector_armed, set_connector_settings,
  connector_debug, _start_armed_connectors

GPS baud constants (_GPS_BYTES_PER_SENTENCE, _GPS_SENTENCES_PER_FIX,
_BAUD_WARN_FRACTION) moved onto HardwareGlue (only used by
_build_serial_gps).

Connector private state: _connector_grants is initialised in
Runtime.__init__ and accessed directly by tests (rt._connector_grants),
so it stays on Runtime; HardwareGlue accesses it via self._rt._connector_grants.

Runtime shims kept for every method: all external callers in server.py and
tests use the Runtime-level API (patch.object(Runtime, "_build_serial_motor",
...) etc.) -- one-line shims preserve exact signatures with zero callsite
changes required.

HardwareGlue is constructed in Runtime.__init__ BEFORE _construct_devices
so the serial-builder shims are ready when DeviceManager._construct_devices
calls them at startup.

app.py -508 lines; hardware_glue.py +613 lines.
All 2137 tests green; ruff clean; no import cycle.
